app/requirements.txt
sqlalchemy
pandas
python-dateutil
mysqlclient
Flask
